adding javadoc and rename AbstractStringLiteral
