+++ src/main/java/fr/inria/corese/core/next/impl/common/util/literal/CoreDatatypeHelper.java
